% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concept_ancestors}
\alias{concept_ancestors}
\alias{concept_descendants}
\title{Fetch active ancestors/descendants of one or more concepts}
\usage{
concept_ancestors(
  conceptIds,
  include_self = FALSE,
  encoding = "UTF-8",
  silent = FALSE,
  ...
)

concept_descendants(
  conceptIds,
  include_self = FALSE,
  encoding = "UTF-8",
  silent = FALSE,
  ...
)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{include_self}{a logical vector indicating whether the \code{conceptIds}
should be included in the results. The default is \code{TRUE}. If a single
value is provided, it will be recycled.}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{...}{other valid arguments to function \code{\link{api_concepts}},
for instance \code{endpoint}, \code{branch} or \code{limit}.}
}
\value{
a named list of data frames
}
\description{
Wrapper functions of \code{\link{api_concepts}} that fetch ancestors
or descendants of one or several concept identifiers using full SNOMED CT
inference.

Note: these functions can only fetch active concepts.
}
\section{Disclaimer}{

In order to use SNOMED CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. See details on \link{snomedizer}.
}

\examples{
pneumonia_ancestors <- concept_ancestors(conceptIds = "233604007")
# This will trigger a warning using the default limit set by snomedizer_options_get("limit")
pneumonia_concepts <- concept_descendants(conceptIds = "233604007")
# Raising the limit
pneumonia_concepts <- concept_descendants(conceptIds = "233604007", limit = 300)
head(pneumonia_concepts$`233604007`)
}
\seealso{
Other wrapper: 
\code{\link{concept_descriptions}()},
\code{\link{concept_find}()},
\code{\link{release_version}()}
}
\concept{wrapper}
