% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concept_descriptions}
\alias{concept_descriptions}
\title{Fetch descriptions of one or more concepts}
\usage{
concept_descriptions(
  conceptIds,
  encoding = "UTF-8",
  silent = FALSE,
  limit = snomedizer_options_get("limit"),
  ...
)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{encoding}{HTTP charset parameter to use. Default is \code{"UTF-8"}.}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}.}

\item{limit}{a positive integer for the maximum number of results to return.
See \code{\link{snomedizer_options}}. The maximum limit on public endpoints
is 10,000.}

\item{...}{other optional arguments listed in \code{\link{api_operations}}, such as
\code{endpoint}, \code{branch} or \code{limit}}
}
\value{
a named list of data frames sorted by \code{conceptIds}
}
\description{
This function is a wrapper of \code{\link{api_descriptions}} that
fetches description of one or several concept identifiers.
}
\section{Note}{

Duplicate \code{conceptIds} will be removed.
}

\section{Disclaimer}{

In order to use SNOMED CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. See details on \link{snomedizer}.
}

\examples{
pneumonia_descriptions <- concept_descriptions(conceptIds = "233604007")
str(pneumonia_descriptions)
}
\seealso{
Other wrapper: 
\code{\link{concept_ancestors}()},
\code{\link{concept_find}()},
\code{\link{release_version}()}
}
\concept{wrapper}
