% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concept_find}
\alias{concept_find}
\title{Find SNOMED CT concepts}
\usage{
concept_find(
  term = NULL,
  conceptIds = NULL,
  ecl = NULL,
  activeFilter = TRUE,
  encoding = "UTF-8",
  silent = FALSE,
  limit = snomedizer_options_get("limit"),
  ...
)
}
\arguments{
\item{term}{a string containing search terms}

\item{conceptIds}{a character vector of one or more SNOMED CT codes
to search}

\item{ecl}{a character expression constraint query (with full inference). Consult the
\href{http://snomed.org/ecl}{Expression Constraint Language guide}
for more detail.}

\item{activeFilter}{whether to restrict results to active concepts. Default is `TRUE`.
Consult the \href{http://snomed.org/gl}{SNOMED glossary} for more detail.}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{limit}{a positive integer for the maximum number of results to return.
See \code{\link{snomedizer_options}}. The maximum limit on public endpoints
is 10,000.}

\item{...}{other optional arguments listed in \code{\link{api_operations}}, such as
\code{endpoint}, \code{branch} or \code{limit}}
}
\value{
a data frame
}
\description{
A wrapper function for \code{\link{api_concepts}} searching for a
term or a list of concepts
}
\section{Disclaimer}{

In order to use SNOMED CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. See details on \link{snomedizer}.
}

\examples{
# Free text search
str(concept_find("asthma"))

# Retrieve multiple concepts
concept_find(conceptIds =  c("233604007", "68566005"))

# Use the SNOMED CT Expression Constraint Language
concept_find(
  ecl = paste(
    "<! 68566005 | Urinary tract infectious disease (disorder) |",
    "AND",
    "< 87628006 | Bacterial infectious disease (disorder) |"
    )
  )
}
\seealso{
Other wrapper: 
\code{\link{concept_ancestors}()},
\code{\link{concept_descriptions}()},
\code{\link{release_version}()}
}
\concept{wrapper}
