% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concept_is}
\alias{concept_is}
\title{Determine whether one or more concept are subtypes of a target set of concepts}
\usage{
concept_is(
  concept_ids,
  target_ecl,
  silent = FALSE,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  encoding = "UTF-8"
)
}
\arguments{
\item{concept_ids}{character vector of identifiers of concepts to be analysed}

\item{target_ecl}{character ECL expression defining the target set of concepts}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{endpoint}{URL of a SNOMED CT Terminology Server REST API endpoint.
See \code{\link{snomedizer_options}}.}

\item{branch}{a string for the name of the API endpoint branch to use (most
commonly \code{"MAIN"}). See \code{\link{snomedizer_options}}.}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}
}
\value{
a logical vector with the same order as \code{concept_ids} indicating
whether each concept is included in the target set or not. \code{NA} is returned
when \code{concept_ids} are not present on the target \code{branch}, or in case
of a REST error caused by invalid \code{concept_ids} or \code{target_ecl}.
}
\description{
This function looks for \code{116680003 | Is a (attribute) |}
relationships. It checks a vector of \code{concept_ids} identifiers
against a target set defined by an ECL expression \code{target_ecl}.
It returns \code{TRUE} if
the concept is a subtype of any concept in the target set,
\code{FALSE} if it does not, or
\code{NA} if it is not found in the branch.
}
\examples{
concept_is(
  concept_ids = "16227691000119107",  # Post-surgical excision site
  target_ecl = "123037004"            # Body structure
)
concept_is(
  concept_ids = "48800003",           # Ear lobule structure
  target_ecl = "233604007"            # Pneumonia
)
concept_is(
  concept_ids = "39732311000001104",  # Medical product only found UK Edition
  target_ecl = "27658006"             # Product containing amoxicillin
)
concept_is(
  concept_ids = "233604007",          # Pneumonia
  target_ecl = "<<64572001 :
         116676008 = <<409774005"     # Disorder with inflammation as associated morphology
)
}
\seealso{
\href{https://confluence.ihtsdotools.org/display/DOCECL/Appendix+D+-+ECL+Quick+reference}{ECL quick reference table by SNOMED International}
}
