% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concept_map}
\alias{concept_map}
\title{Map SNOMED CT concepts to other terminology or code systems}
\usage{
concept_map(
  concept_ids = NULL,
  target_code = NULL,
  map_refset_id = "447562003",
  active = TRUE,
  encoding = "UTF-8",
  silent = FALSE,
  limit = snomedizer_options_get("limit"),
  ...
)
}
\arguments{
\item{concept_ids}{an optional character vector of one or more SNOMED CT concept
identifiers to be mapped}

\item{target_code}{an optional character code designated the concept code in
the other terminology or code system}

\item{map_refset_id}{character identifier of a SNOMED CT Map Reference Set.
The default is \code{"447562003"} for the ICD-10 Map Reference Set.}

\item{active}{whether to restrict results to active concepts. Default is \code{TRUE}.}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{limit}{a positive integer for the maximum number of results to return.
See \code{\link{snomedizer_options}}. The maximum limit on public endpoints
is 10,000.}

\item{...}{other optional arguments listed in \code{\link{api_operations}}, such as
\code{endpoint}, \code{branch} or \code{limit}}
}
\value{
a data frame of SNOMED CT concepts mapped to another code system
}
\description{
A wrapper function for the \code{\link{api_refset_members}()} function
to query Map Reference Sets, in particular the map to the World Health Organisation
International Classification of Diseases 10th Revision (ICD-10).
}
\examples{
# find SNOMED CT codes corresponding to ICD-10 code N39.0 urinary tract infections
uti_concepts <- concept_map(target_code = "N39.0")
str(dplyr::select(uti_concepts,
                  referencedComponentId,
                  referencedComponent.pt.term,
                  additionalFields.mapTarget,
                  additionalFields.mapAdvice))

# map SNOMED CT codes to ICD-10
map_icd10 <- concept_map(concept_ids = c("431308006", "312124009", "53084003"))
dplyr::select(map_icd10,
              referencedComponentId,
              referencedComponent.pt.term,
              additionalFields.mapTarget,
              additionalFields.mapAdvice)
}
\seealso{
World Health Organisation \href{https://icd.who.int/browse10/2016/en}{International Classification of Diseases 10th Revision}

SNOMED International \href{https://confluence.ihtsdotools.org/display/DOCRELFMT/5.2.10+Complex+and+Extended+Map+Reference+Sets}{Map Reference Sets}

SNOMED International \href{http://snomed.org/icd10map}{ICD-10 Mapping Technical Guide}
}
