% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_ascendants}
\alias{concepts_ascendants}
\title{Fetch ascendants of one or more concepts}
\usage{
concepts_ascendants(
  conceptIds,
  direct_ascendants = FALSE,
  activeFilter = TRUE,
  encoding = "UTF-8",
  silent = FALSE,
  ...
)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{direct_ascendants}{a logical vector indicating whether to fetch
direct ascendants (parents) of the \code{conceptIds} exclusively or all
ascendants (including parents). The default is \code{FALSE}. If a single
value is provided, it will be recycled.}

\item{activeFilter}{a logical vector indicating whether to fetch active
ascendant concepts exclusively. The default is \code{TRUE}. If a single
value is provided, it will be recycled.}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{...}{other valid arguments to function \code{\link{api_concepts}},
for instance \code{endpoint}, \code{branch} or \code{limit}.}
}
\value{
a named list of data frames
}
\description{
This function is a wrapper of \code{\link{api_concepts}} that
fetches ascendants of one or several concept identifiers using full SNOMED CT
inference. The search can be restricted to parent concepts using the
\code{direct_ascendants} argument.
}
\section{Disclaimer}{

In order to use SNOMED CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. See details on \link{snomedizer}.
}

\examples{
# This will trigger a warning using the default limit set by snomedizer_options_get("limit")
pneumonia_concepts <- concepts_ascendants(conceptIds = "233604007")
}
\seealso{
Other wrapper: 
\code{\link{concepts_descendants}()},
\code{\link{concepts_descriptions}()},
\code{\link{concepts_find}()},
\code{\link{release_version}()}
}
\concept{wrapper}
