% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_descendants}
\alias{concepts_descendants}
\title{Fetch descendants of one or more concepts}
\usage{
concepts_descendants(
  conceptIds,
  direct_descendants = FALSE,
  activeFilter = TRUE,
  ...
)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{direct_descendants}{a logical vector indicating whether to fetch
direct descendants (children) of the \code{conceptIds} exclusively or all
descendants. The default is \code{FALSE}. If a single
value is provided, it will be recycled.}

\item{activeFilter}{a logical vector indicating whether to fetch active
descendant concepts exclusively. The default is \code{TRUE}. If a single
value is provided, it will be recycled.}

\item{...}{other valid arguments to function \code{\link{api_concepts}},
for instance \code{endpoint}, \code{branch} or \code{limit}.}
}
\value{
a named list of data frames
}
\description{
This function is a wrapper of \code{\link{api_concepts}} that
fetches descendants of one or several concept identifiers using full SNOMED-CT
inference. The search can be restricted to children concepts using the
\code{direct_descendants} argument.
}
\examples{
# This will trigger a warning using the default limit set by snomedizer_options_get("limit")
concepts_descendants(conceptIds = c("233604007", "68566005"))
# Raising the limit
concepts_descendants(conceptIds = c("233604007", "68566005"), limit = 300)
}
