% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_descriptions}
\alias{concepts_descriptions}
\title{Fetch descriptions of one or more concepts}
\usage{
concepts_descriptions(conceptIds, encoding = "UTF-8", ...)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}

\item{...}{other optional arguments listed in \code{\link{api_operations}}, such as
\code{endpoint}, \code{branch} or \code{limit}}
}
\value{
a named list of data frames sorted by \code{conceptIds}
}
\description{
This function is a wrapper of \code{\link{api_descriptions}} that
fetches description of one or several concept identifiers.
}
\section{Note}{

Duplicate \code{conceptIds} will be removed.
}

\section{Disclaimer}{

In order to use SNOMED CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. See details on \link{snomedizer}.
}

\examples{
pneumonia_descriptions <- concepts_descriptions(conceptIds = "233604007")
str(pneumonia_descriptions)
}
\seealso{
Other wrapper: 
\code{\link{concepts_descendants}()},
\code{\link{concepts_find}()},
\code{\link{release_version}()}
}
\concept{wrapper}
