% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_find}
\alias{concepts_find}
\title{Find SNOMED-CT concepts}
\usage{
concepts_find(
  term = NULL,
  conceptIds = NULL,
  ecl = NULL,
  activeFilter = TRUE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{term}{a string containing search terms}

\item{conceptIds}{a character vector of one or more SNOMED-CT codes
to search}

\item{ecl}{a character expression constraint query (with full inference). Consult the
\href{http://snomed.org/ecl}{Expression Constraint Language guide}
for more detail.}

\item{activeFilter}{whether to restrict results to active concepts. Default is `TRUE`.
Consult the \href{http://snomed.org/gl}{SNOMED glossary} for more detail.}

\item{silent}{whether to hide warnings. Default is `FALSE`}

\item{...}{other optional arguments listed in \code{\link{api_operations}}, such as
\code{endpoint}, \code{branch} or \code{limit}}
}
\value{
a data frame
}
\description{
A wrapper function for \code{\link{api_concepts}} searching for a
term or a list of concepts
}
\examples{
# Free text search
concepts_find("asthma")

# Retrieve multiple concepts
concepts_find(conceptIds =  c("233604007", "68566005"))

# Use the SNOMED-CT Expression Constraint Language
concepts_find(
  ecl = paste(
    "<! 68566005 | Urinary tract infectious disease (disorder) |",
    "AND",
    "< 87628006 | Bacterial infectious disease (disorder) |"
    )
  )
}
\concept{wrapper}
