% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_included_in}
\alias{concepts_included_in}
\title{Determine where a concept of interest belongs to a target set of concepts}
\usage{
concepts_included_in(
  concept_ids,
  target_ecl,
  silent = FALSE,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch")
)
}
\arguments{
\item{concept_ids}{character vector of identifiers of domain concepts to be analysed}

\item{target_ecl}{character ECL expression defining the target set of concepts}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{endpoint}{URL of a SNOMED CT Terminology Server REST API endpoint.
See \code{\link{snomedizer_options}}.}

\item{branch}{a string for the name of the API endpoint branch to use (most
commonly \code{"MAIN"}). See \code{\link{snomedizer_options}}.}
}
\value{
a logical vector with the same order as \code{concept_ids} indicating
whether each concept is included in the target set or not. \code{NA} denotes
a REST error.
}
\description{
This function examines a vector of \code{concept_ids} identifiers
against a target set defined by an ECL expression
}
\examples{
concepts_included_in(
  concept_ids = "16227691000119107 | Post-surgical excision site",
  target_ecl = "123037004 | Body structure"
)
concepts_included_in(
  concept_ids = "48800003 | Ear lobule structure (body structure) |",
  target_ecl = "233604007 | Pneumonia (disorder) |"
)
}
\seealso{
\href{https://confluence.ihtsdotools.org/display/DOCECL/Appendix+D+-+ECL+Quick+reference}{ECL quick reference table by SNOMED International}
}
