% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_map}
\alias{concepts_map}
\title{Map SNOMED CT concepts to other terminology or code systems}
\usage{
concepts_map(
  concept_ids = NULL,
  target_code = NULL,
  map_refset_id = "447562003",
  active = TRUE,
  encoding = "UTF-8",
  silent = FALSE,
  ...
)
}
\arguments{
\item{concept_ids}{an optional character vector of one or more SNOMED CT concept
identifiers to be mapped}

\item{target_code}{an optional character code designated the concept code in
the other terminology or code system}

\item{map_refset_id}{character identifier of a SNOMED CT Map Reference Set.
The default is \code{"447562003"}}

\item{active}{whether to restrict results to active concepts. Default is \code{TRUE}.}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}

\item{silent}{whether to hide progress bar. Default is \code{FALSE}}

\item{...}{other optional arguments listed in \code{\link{api_operations}}, such as
\code{endpoint}, \code{branch} or \code{limit}}
}
\value{
a data frame of SNOMED CT concepts mapped to another code system
}
\description{
A wrapper function for the \code{\link{api_refset_members}()} function.
}
\examples{
# find SNOMED CT codes corresponding to urinary tract infections N39.0
uti_concepts <- concepts_map(target_code = "N39.0")
dplyr::select(uti_concepts,
              referencedComponentId,
              referencedComponent.pt.term,
              additionalFields.mapTarget,
              additionalFields.mapAdvice)

# map SNOMED CT codes to ICD=10 codes
map_icd10 <- concepts_map(concept_ids = c("431308006", "312124009", "53084003"))
dplyr::select(map_icd10,
              referencedComponentId,
              referencedComponent.pt.term,
              additionalFields.mapTarget,
              additionalFields.mapAdvice)
}
