% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.catch404}
\alias{.catch404}
\title{Catch 404 errors in a server request}
\usage{
.catch404(x)
}
\arguments{
\item{x}{a \code{\link{httr}{response}} object from a server request}
}
\description{
Catch 404 errors in a server request
}
\keyword{internal}
