% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.catch_http_error}
\alias{.catch_http_error}
\title{Catch HTTP errors in a server request}
\usage{
.catch_http_error(x)
}
\arguments{
\item{x}{a \code{\link{httr}{response}} object from a server request}
}
\description{
Catch HTTP errors in a server request
}
\keyword{internal}
