% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_rest_query_length1}
\alias{.check_rest_query_length1}
\title{Check that all REST query parameters have length 1}
\usage{
.check_rest_query_length1(rest_url)
}
\arguments{
\item{rest_url}{a list of class \code{url} generated by \code{\link[httr]{parse_url}()}
and containing a \code{query} list}
}
\description{
Check that all REST query parameters have length 1
}
\keyword{internal}
