% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.concatenate_array_parameter}
\alias{.concatenate_array_parameter}
\title{Concatenated REST query array parameters}
\usage{
.concatenate_array_parameter(param)
}
\arguments{
\item{param}{a character vector of values to concatenate}
}
\value{
a character string
}
\description{
Concatenated REST query array parameters
}
\keyword{internal}
