% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_indications.R
\docType{data}
\name{drug_indications}
\alias{drug_indications}
\title{Sample of antibiotic prescription records}
\format{
An object of class \code{data.frame} with 263 rows and 2 columns.
}
\usage{
drug_indications
}
\description{
Dataset containing 403 records of free-text therapy indications
filled in by doctors when prescription antibiotics
}
\details{
A data frame with one row per therapy indication:
\describe{
   \item{\code{prescription_id}}{unique identifier for the record}
   \item{\code{indication}}{character vector of free-text therapy indication
   for the antibiotic prescription (eg \code{"urinary tract infection"})}
}
}
\keyword{datasets}
