% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{release_version}
\alias{release_version}
\title{Fetch endpoint release version}
\usage{
release_version(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch")
)
}
\arguments{
\item{endpoint}{the URL of a SNOMED CT Terminology Server REST API endpoint.
See \code{\link{snomedizer_options}}.}

\item{branch}{a string for the name of the API endpoint branch to use (most
commonly \code{"MAIN"}). See \code{\link{snomedizer_options}}.}
}
\value{
a list containing two character strings: \code{rf2_date}
(YYYYMMDD release date) and \code{rf2_month_year} (month and year string)
}
\description{
Fetch endpoint release version
}
