% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{result_completeness}
\alias{result_completeness}
\title{Check that results are complete}
\usage{
result_completeness(x, silent = FALSE)
}
\arguments{
\item{x}{an `httr` \code{\link[httr]{response}()} object produce by an
\code{\link{api_operations}} function.}

\item{silent}{whether to display warnings (default is `FALSE`)}
}
\value{
a boolean indicating whether the set of results obtained is
the complete set of results on the server.
}
\description{
Check whether the server request returned all the results,
i.e. whether the `limit` < results `total`.
}
\examples{
result_completeness(api_find_concepts(term = "pneumonia", limit = 10))
}
