% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{result_flatten}
\alias{result_flatten}
\title{Flatten results from a server request}
\usage{
result_flatten(x)
}
\arguments{
\item{x}{an `httr` \code{\link[httr]{response}()} object}
}
\value{
a data frame
}
\description{
A function to \code{\link[jsonlite]{flatten}}
nested server results from an \code{\link{api_operations}}
 into a single non-nested data frame
}
\examples{
result_flatten(
   api_find_concepts(term = "pneumonia",
   activeFilter = TRUE,
   limit = 10))
}
