% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snomed_endpoint_test}
\alias{snomed_endpoint_test}
\title{Test a SNOMED CT endpoint}
\usage{
snomed_endpoint_test(endpoint, branch)
}
\arguments{
\item{endpoint}{URL of a SNOMED CT Terminology Server REST API endpoint}

\item{branch}{a character string of a branch name}
}
\value{
a boolean indicating whether the endpoint passed the test
}
\description{
Test a SNOMED CT endpoint
}
\examples{
snomed_endpoint_test(
  endpoint = "https://snowstorm.ihtsdotools.org/snowstorm/snomed-ct",
  branch = "MAIN"
)
}
\concept{utilities#'}
