% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snomed_public_endpoint_list}
\alias{snomed_public_endpoint_list}
\title{List of public SNOMED CT endpoints}
\usage{
snomed_public_endpoint_list()
}
\value{
a vector of URLs
}
\description{
List a range of know public SNOMED CT Terminology Server
REST API endpoint. To select a currently active endpoint, see
\code{\link{snomed_public_endpoint_suggest}()}
}
\seealso{
Other utilities: 
\code{\link{result_completeness}()},
\code{\link{result_flatten}()},
\code{\link{snomed_public_endpoint_suggest}()},
\code{\link{snomedizer_options}},
\code{\link{snomedizer_version_compatibility}()}
}
\concept{utilities}
