% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snomedizer-package.R
\docType{package}
\name{snomedizer-package}
\alias{snomedizer}
\alias{snomedizer-package}
\title{snomedizer: R Interface to the SNOMED CT Terminology Server REST API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Interrogate the SNOMED CT clinical ontology using the 
   SNOMED International Terminology Server REST API <https://github.com/IHTSDO/snowstorm>.
}
\section{SNOMED CT licensing}{


In order to use SNOMED CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. Although SNOMED International allows access to the
SNOMED CT terminology server hosted by IHSTO to anyone for reference purpose, uses for data
analysis or health care production systems is subject to other licences.

SNOMED International allows access to public SNOMED CT terminology servers to anyone
for strict 'reference purposes' under the \href{https://browser.ihtsdotools.org/}{SNOMED
International SNOMED CT Browser License Agreement} (see section below).

Use of SNOMED CT terminology for data analysis or health care production systems is
subject to other licences. Some users are eligible for free licences:
\itemize{
  \item{UK-based users can obtain a licence free of charge on the
  \href{https://isd.digital.nhs.uk/trud3}{NHS TRUD website}}
  \item{residents of other Member Countries and low-income countries are also eligible.
  More information can be found on
  \href{https://www.snomed.org/snomed-ct/get-snomed}{SNOMED International's website}.}
}
}

\section{Licensing for reference use}{


\code{snomedizer} contains examples and documentation that makes uses of an IHSTO public
API endpoint restricted to reference uses such as testing or browsing the terminology.
It is subject to the
\href{https://browser.ihtsdotools.org/}{SNOMED International Browser License Agreement}
reproduced below:

\enumerate{
\item The meaning of the terms "Affiliate", or "Data Analysis System", "Data Creation System",
"Derivative", "End User", "Extension", "Member", "Non-Member Territory", "SNOMED CT"
and "SNOMED CT Content" are as defined in the SNOMED International Affiliate License Agreement
(see \url{http://snomed.org/license-affiliate}).

\item Information about Affiliate Licensing is available at \url{https://www.snomed.org/snomed-ct/get-snomed}.
Individuals or organizations wishing to register as SNOMED International Affiliates can register at
\url{mlds.ihtsdotools.org}, subject to acceptance of the Affiliate License Agreement (see
\url{http://snomed.org/license-affiliate}).

\item The current list of SNOMED International Member Territories can be viewed at \url{www.snomed.org/members}.
Countries not included in that list are "Non-Member Territories".

\item End Users, that do not hold an SNOMED International Affiliate License, may access SNOMED CT®
using SNOMED International SNOMED CT Browser subject to acceptance of and adherence to the
following sub-license limitations: \itemize{
  \item a) The sub-licensee is only permitted to access SNOMED CT® using this software (or service) for the purpose of exploring and evaluating the terminology.
  \item b) The sub-licensee is not permitted the use of this software as part of a system that constitutes a SNOMED CT "Data Creation System" or "Data Analysis System", as defined in the SNOMED International Affiliate License. This means that the sub-licensee must not use SNOMED International SNOMED CT Browser to add or copy SNOMED CT identifiers into any type of record system, database or document.
  \item c) The sub-licensee is not permitted to translate or modify SNOMED CT Content or Derivatives.
  \item d) The sub-licensee is not permitted to distribute or share SNOMED CT Content or Derivatives.
}

\item SNOMED International Affiliates may use SNOMED International SNOMED CT Browser as part of a
"Data Creation System" or "Data Analysis System" subject to the following conditions: \itemize{
  \item a) The SNOMED International Affiliate, using SNOMED International SNOMED CT Browser must accept full responsibility for any reporting and fees due for use or deployment of such a system in a Non-Member Territory.
  \item b) The SNOMED International Affiliate must not use SNOMED International SNOMED CT Browser to access or interact with SNOMED CT in any way that is not permitted by the Affiliate License Agreement.
  \item c) In the event of termination of the Affiliate License Agreement, the use of SNOMED International SNOMED CT Browser will be subject to the End User limitations noted in 4.
  }
}
}

\section{snomedizer licence}{

Copyright © 2020 University College London

This program is free software; you can redistribute it and/or modify it
under the terms of the GNU General Public License as published by the Free
Software Foundation; either version 3 of the License, or (at your option)
any later version.

This program is distributed in the hope that it will be useful, but WITHOUT
ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramses-antibiotics/snomedizer}
  \item Report bugs at \url{https://github.com/ramses-antibiotics/snomedizer/issues}
}

}
\author{
\strong{Maintainer}: Peter Dutey-Magni \email{p.dutey-magni@ucl.ac.uk} (\href{https://orcid.org/0000-0002-8942-9836}{ORCID}) [researcher]

Other contributors:
\itemize{
  \item Anika Cawthorn \email{a.cawthorn@ucl.ac.uk} (\href{https://orcid.org/0000-0002-2438-7495}{ORCID}) [reviewer, researcher]
  \item University College London [copyright holder]
}

}
\keyword{internal}
