% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snomedizer_options}
\alias{snomedizer_options}
\alias{snomedizer_options_get}
\alias{snomedizer_options_set}
\title{Default \code{snomedizer} options}
\usage{
snomedizer_options_get(option.name = NULL)

snomedizer_options_set(endpoint = NULL, branch = NULL, limit = NULL)
}
\arguments{
\item{option.name}{the name of an option to return. If NULL (the default),
\code{snomedizer_options_get()} returns a list of all options.}

\item{endpoint}{URL of a SNOMED CT Terminology Server REST API endpoint.
When the package is loaded, this option is set to the current environment variable
\code{SNOMEDIZER_ENDPOINT} if it exists, or to
\code{\link{snomed_public_endpoint_suggest}()} otherwise.}

\item{branch}{string for the branch to use on the SNOMED CT Terminology Server REST API
endpoint. Set to `"MAIN"` by default when the package is loaded.}

\item{limit}{integer for the maximum number of results to return. Set to 50 by default
when the package is loaded.}
}
\value{
The factory setting of the target API parameter.
}
\description{
Functions to get and set \code{snomedizer} default endpoint and other options
}
\details{
The SNOMED REST API requires several parameters which are
set to default values when this package is launched:
\itemize{
   \item{endpoint} this is the address of the SNOWSTORM terminology server
   to be used. When \code{snomedizer} is loaded, it is set to the current
   environment variable \code{SNOMEDIZER_ENDPOINT}. If no such
    the SNOMED-CT version maintained by the official
        SNOWSTORM server.
   \item{branch} The default is "MAIN/SNOMEDCT-GB", the most
        up-to-date edition of SNOMED-CT UK Core Edition.
   \item{limit} an integer stating the the maximum number of results fetched. This is set to 50 The default
        is 50.
}
}
\examples{
snomedizer_options_get()
}
\seealso{
To learn how to set environment variables in `.Rprofile` or `.Renviron`, see
\link[base]{Startup}

Other utilities: 
\code{\link{result_completeness}()},
\code{\link{result_flatten}()},
\code{\link{snomed_public_endpoint_suggest}()}
}
\concept{utilities}
