% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snomedizer_version_compatibility}
\alias{snomedizer_version_compatibility}
\title{Verify the endpoint compatibility with snomedizer}
\usage{
snomedizer_version_compatibility(
  endpoint = snomedizer_options_get("endpoint"),
  silent = FALSE
)
}
\arguments{
\item{endpoint}{URL of a SNOMED CT Terminology Server REST API endpoint}

\item{silent}{whether to hide warnings. Default is `FALSE`}
}
\value{
a logical value indicating whether the endpoint version is supported
}
\description{
This function compares the SNOMED CT terminology server endpoint version
with `snomedizer`'s supported version. The SNOMED CT terminology server API
is continuously developed and may introduce breaking changes in REST operations
and parameters.
}
\seealso{
Other utilities: 
\code{\link{result_completeness}()},
\code{\link{result_flatten}()},
\code{\link{snomed_public_endpoint_list}()},
\code{\link{snomed_public_endpoint_suggest}()},
\code{\link{snomedizer_options}}
}
\concept{utilities}
