% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{api_operations}
\alias{api_operations}
\alias{api_find_concept}
\alias{api_find_concepts}
\alias{api_all_branches}
\alias{api_branch}
\alias{api_branch_descendants}
\alias{api_descriptions}
\alias{api_version}
\alias{api_browser_concepts}
\alias{api_browser_concept_ancestors}
\title{SNOMED CT Terminology Server REST API operations}
\usage{
api_find_concept(
  conceptId,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  accept_language = snomedizer_options_get("language"),
  catch404 = TRUE
)

api_find_concepts(
  term = NULL,
  conceptIds = NULL,
  activeFilter = NULL,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  limit = snomedizer_options_get("limit"),
  accept_language = snomedizer_options_get("language"),
  offset = 0,
  catch404 = TRUE,
  ...
)

api_all_branches(
  endpoint = snomedizer_options_get("endpoint"),
  catch404 = TRUE
)

api_branch(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE,
  ...
)

api_branch_descendants(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE,
  ...
)

api_descriptions(
  concept,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  offset = 0,
  limit = snomedizer_options_get("limit"),
  catch404 = TRUE,
  ...
)

api_version(endpoint = snomedizer_options_get("endpoint"), catch404 = TRUE)

api_browser_concepts(
  conceptId,
  descendantCountForm = c(NULL, "inferred", "stated", "additional"),
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  accept_language = snomedizer_options_get("language"),
  catch404 = TRUE
)

api_browser_concept_ancestors(
  conceptId,
  form = c("inferred", "stated", "additional"),
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  accept_language = snomedizer_options_get("language"),
  catch404 = TRUE
)
}
\arguments{
\item{conceptId}{character string of a SNOMED-CT concept id (for example:
\code{"233604007"})}

\item{endpoint}{the URL of a SNOMED CT Terminology Server REST API endpoint.
See \code{\link{snomedizer_options}}.}

\item{branch}{a string for the name of the API endpoint branch to use (most
commonly \code{"MAIN"}). See \code{\link{snomedizer_options}}.}

\item{accept_language}{a string specifying acceptable result languages}

\item{catch404}{whether to display a warning if the API operation returns a
'404 Not Found' status. Default is \code{TRUE}.}

\item{term}{character vector of terms to search}

\item{conceptIds}{a character vector of SNOMED-CT concept ids (for example:
\code{c("233604007", "68566005")})}

\item{activeFilter}{optional boolean: \itemize{
    \item \code{TRUE} returns only active terminology
    \item \code{FALSE} returns only inactive terminology
    \item \code{NULL} (the default) returns both active and inactive terminology
}}

\item{limit}{integer for the maximum number of results to return.
See \code{\link{snomedizer_options}}.}

\item{offset}{an integer indicating the number of results to skip}

\item{...}{other REST API parameters}

\item{concept}{character string of a SNOMED-CT concept id (for example:
\code{"233604007"})}

\item{descendantCountForm}{a character string indicating whether to report
the count of descendant concepts based on stated or inferred relationships.
Must be one of \code{"inferred"}, \code{"stated"}, or \code{"additional"}.
Default is \code{NULL} for no descendant count reported.}

\item{form}{a character string indicating which ancestors/descendants to
extract based on stated or inferred relationships. Must be one of
\code{"inferred"} (default), \code{"stated"}, or \code{"additional"}.}
}
\value{
An \code{httr} \code{\link[httr]{response}()} object.
}
\description{
Low-level wrapper functions based on \link{httr} for interfacing
with the operations built in the \href{https://github.com/IHTSDO/snowstorm}{Snowstorm API}
}
\examples{
# look up the pneumonia concept
api_find_concept(conceptId = "233604007")
api_find_concepts(term = "pneumonia")
api_find_concepts(conceptIds = c("233604007", "68566005"))

# get the content of the server request
httr::content(api_find_concepts(term = "pneumonia"))
}
