% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{concepts_descendants}
\alias{concepts_descendants}
\title{Fetch descendants of one or more concepts}
\usage{
concepts_descendants(
  conceptIds,
  direct_descendants = FALSE,
  activeFilter = TRUE,
  ...
)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{direct_descendants}{a logical vector indicating whether to fetch
direct descendants (children) of the \code{conceptIds} exclusively or all
descendants. The default is \code{FALSE}. If a single
value is provided, it will be recycled.}

\item{activeFilter}{a logical vector indicating whether to fetch active
descendant concepts exclusively. The default is \code{TRUE}. If a single
value is provided, it will be recycled.}

\item{...}{other valid arguments to function \code{\link{api_concepts}},
for instance \code{endpoint}, \code{branch} or \code{limit}.}
}
\value{
a named list of data frames
}
\description{
This function is a wrapper of \code{\link{api_concepts}} that
fetches descendants of one or several concept identifiers using full SNOMED-CT
inference. The search can be restricted to children concepts using the
\code{direct_descendants} argument.
}
\section{Disclaimer}{

In order to use SNOMED-CT, a licence is required which depends both on the country you are
based in, and the purpose of your work. See details on \link{snomedizer}.
}

\examples{
# This will trigger a warning using the default limit set by snomedizer_options_get("limit")
pneumonia_concepts <- concepts_descendants(conceptIds = "233604007")
# Raising the limit
pneumonia_concepts <- concepts_descendants(conceptIds = "233604007", limit = 300)
head(pneumonia_concepts$`233604007`)
}
