% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{release_version}
\alias{release_version}
\title{Fetch SNOMED CT RF2 release version}
\usage{
release_version(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch")
)
}
\arguments{
\item{endpoint}{the URL of a SNOMED CT Terminology Server REST API endpoint.
See \code{\link{snomedizer_options}}.}

\item{branch}{a string for the name of the API endpoint branch to use (most
commonly \code{"MAIN"}). See \code{\link{snomedizer_options}}.}
}
\value{
a list containing two character strings: \code{rf2_date}
(YYYYMMDD release date) and \code{rf2_month_year} (month and year string)
}
\description{
Provides the date of the release of the specified endpoint
and branch. SNOMED CT is currently released released twice a year
on 31 January and 31 July in a format known as Release Format 2 (RF2).
}
\references{
\href{SNOMED CT Release File Specifications}{http://snomed.org/rfs}
}
\seealso{
Other wrapper: 
\code{\link{concept_ancestors}()},
\code{\link{concept_descriptions}()},
\code{\link{concept_find}()}
}
\concept{wrapper}
