% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{result_completeness}
\alias{result_completeness}
\title{Check that results are complete}
\usage{
result_completeness(x, silent = FALSE)
}
\arguments{
\item{x}{an `httr` \code{\link[httr]{response}()} object produce by an
\code{\link{api_operations}} function.}

\item{silent}{whether to display warnings (default is `FALSE`)}
}
\value{
a boolean indicating whether the set of results obtained is
the complete set of results on the server. If the REST request failed
(eg: 404 error), the function returns \code{NULL}
}
\description{
Check whether the server request returned all the results,
i.e. whether the total number of results is smaller or equal to the
request \code{limit} parameter.
}
\examples{
result_completeness(api_concepts(term = "pneumonia", limit = 10))
}
\seealso{
Other utilities: 
\code{\link{result_flatten}()},
\code{\link{snomed_public_endpoint_list}()},
\code{\link{snomed_public_endpoint_suggest}()},
\code{\link{snomedizer_options}},
\code{\link{snomedizer_version_compatibility}()}
}
\concept{utilities}
