% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{result_flatten}
\alias{result_flatten}
\title{Flatten results from a server request}
\usage{
result_flatten(x, encoding = "UTF-8")
}
\arguments{
\item{x}{an `httr` \code{\link[httr]{response}()} object}

\item{encoding}{HTTP charset parameter to use (default is \code{"UTF-8"})}
}
\value{
a data frame
}
\description{
A function to \code{\link[jsonlite]{flatten}}
nested server results from an \code{\link{api_operations}}
 into a single non-nested data frame
}
\examples{
flattened_results <- result_flatten(
   api_concepts(term = "pneumonia",
   activeFilter = TRUE,
   limit = 10))
str(flattened_results)
}
\seealso{
Other utilities: 
\code{\link{result_completeness}()},
\code{\link{snomed_public_endpoint_list}()},
\code{\link{snomed_public_endpoint_suggest}()},
\code{\link{snomedizer_options}},
\code{\link{snomedizer_version_compatibility}()}
}
\concept{utilities}
