% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snomed_public_endpoint_suggest}
\alias{snomed_public_endpoint_suggest}
\title{Find a public SNOMED CT endpoint}
\usage{
snomed_public_endpoint_suggest()
}
\value{
a string object containing the URL to a responsive SNOMED CT Terminology Server REST API endpoint.
}
\description{
Find a public SNOMED CT endpoint
}
\seealso{
Other utilities: 
\code{\link{result_completeness}()},
\code{\link{result_flatten}()},
\code{\link{snomed_public_endpoint_list}()},
\code{\link{snomedizer_options}},
\code{\link{snomedizer_version_compatibility}()}
}
\concept{utilities}
