% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snomedizer_options}
\alias{snomedizer_options}
\alias{snomedizer_options_get}
\alias{snomedizer_options_set}
\title{Set SNOMED CT endpoint and other \code{snomedizer} options}
\usage{
snomedizer_options_get(option.name = NULL)

snomedizer_options_set(endpoint = NULL, branch = NULL, limit = NULL)
}
\arguments{
\item{option.name}{name of a single option to return: \code{"endpoint"},
\code{"branch"}, or \code{"limit"}. If \code{NULL} (the default),
\code{snomedizer_options_get()} returns a list of all options.}

\item{endpoint}{address of a SNOMED CT Terminology Server REST API endpoint}

\item{branch}{string for the branch name to use on endpoint, for instance
\code{"MAIN"} for the root branch (usually the latest release of
SNOMED CT's International Edition), or \code{"MAIN/2017-07-31"} for a past release.
To obtain a list of all branches available on the current endpoint,
see \code{\link{api_branch}()}}

\item{limit}{integer for the maximum number of results to return.}
}
\value{
The factory setting of the target API parameter.
}
\description{
Functions to get and set \code{snomedizer} default endpoint and other options
}
\section{Default settings and environment variables}{


When loaded, the snomedizer package will look up for settings provided
to the following environment variables:
\describe{
   \item{\code{SNOMEDIZER_ENDPOINT}}{for the \code{endpoint}. If this variable
   is not specified, snomedizer uses
   \code{\link{snomed_public_endpoint_suggest}()} to pick a public endpoint.}
   \item{\code{SNOMEDIZER_BRANCH}}{for the \code{branch}. If this variable is
   not specified, snomedizer chooses branch \code{"MAIN"} by default.}
   \item{\code{SNOMEDIZER_LIMIT}}{for the \code{limit}. If this variable is
   not specified, snomedizer sets the limit to 50 by default.
   The maximum is 10,000.}
}
}

\examples{
snomedizer_options_get()
}
\seealso{
To learn how to set environment variables in `.Rprofile` or `.Renviron`, see
\link[base]{Startup}

Other utilities: 
\code{\link{result_completeness}()},
\code{\link{result_flatten}()},
\code{\link{snomed_public_endpoint_list}()},
\code{\link{snomed_public_endpoint_suggest}()},
\code{\link{snomedizer_version_compatibility}()}
}
\concept{utilities}
